requirements.txt
annotated-types 0.7.0
anyio 4.9.0
blinker 1.9.0
certifi 2025.6.15
charset-normalizer 3.4.2
click 8.1.8
distro 1.9.0
exceptiongroup 1.3.0
Flask 2.3.3
Flask-Cors 4.0.0
gunicorn 21.2.0
h11 0.16.0
httpcore 1.0.9
httpx 0.28.1
idna 3.10
importlib_metadata 8.7.0
itsdangerous 2.2.0
Jinja2 3.1.6
MarkupSafe 3.0.2
openai 1.30.1
packaging 25.0
pydantic 2.11.7
pydantic_core 2.33.2
python-dotenv 1.0.1
rapidfuzz 3.6.2
 8ff5b618021ad6232e52060f8bca3fa754ab8ed7
requests 2.32.4
sniffio 1.3.1
stripe 9.7.0
tqdm 4.67.1
typing-inspection 0.4.1
typing_extensions 4.14.0
urllib3 2.5.0
Werkzeug 3.1.3
zipp 3.23.0
